Add query parameters for appName, appVersion and groupState
Query parameters allow the user to filter the data that is returned by
different queries. This gives more flexibility when creating panels and
data can be queried more efficiently.

For example, the user can now easily request all errors which are open
(where the corresponding error group is open). Or he can filter for a
specific app name and/or app version.